Cache Manager
Added the cache manager which allows any class inheriting from QObject
to be 'cached': I.e all class instances are stored and if the init
arguements of a new instance match those of an existing instance, the
existing instance is returned, rather than a new instance being created.

Useful to allow multiple widgets to talk to manipulate each other.